+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - Fixed crash with broken config
 - Fixed the default for manual clip types being 1000x longer than they were meant to be
 - Localized export grouping modes
+- Localized inverse option
+- Localized export grouping mode option